update project structure + rename files + fix import bug
